💄 update Profile type radio groupd design
